requirements.txt
Flask==2.0.1
numpy==1.21.0
pandas==1.3.0
mysql-connector-python==8.0.25
matplotlib==3.4.2
seaborn==0.11.2
gunicorn==20.1.0  
scikit-learn==0.24.2
